requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.3.0
cryptography==40.0.2
Deprecated==1.2.13
Django==3.2.18
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-filter==23.2
django-oauth-toolkit==2.2.0
django-otp==1.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.6.0
jwcrypto==1.4.2
kombu==5.2.4
Markdown==3.4.3
oauth2-provider==0.0
oauthlib==3.2.2
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psycopg2-binary==2.9.6
pycparser==2.21
pyotp==2.8.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.15.0
